Refatoracao na CustomerController para uso de Service
