+++ SalesProject.Api/Controllers/CustomerController.cs
@@ -2,13 +2,8 @@
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using SalesProject.Api.ViewModels.Customer;
-using SalesProject.Domain.Entities;
-using SalesProject.Domain.Enums;
-using SalesProject.Domain.Interfaces;
-using SalesProject.Domain.Interfaces.Repository;
 using SalesProject.Domain.Interfaces.Service;
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net.Mime;
 
@@ -17,19 +12,15 @@ namespace SalesProject.Api.Controllers
     [ApiController]
     public class CustomerController : ControllerBase
     {
-        private readonly ICustomerRepository _customerRepository;
+        private readonly ICustomerService _customerService;
         private readonly ICnpjApiService _cnpjApiService;
-        private readonly IUnitOfWork _uow;
 
         public CustomerController(
-            ICustomerRepository customerRepository,
-            ICnpjApiService cnpjApiService,
-            IAddressApiService addressApiService,
-            IUnitOfWork uow)
+            ICustomerService customerService,
+            ICnpjApiService cnpjApiService)
         {
-            _customerRepository = customerRepository;
+            _customerService = customerService;
             _cnpjApiService = cnpjApiService;
-            _uow = uow;
         }
 
         /// <summary>
@@ -42,8 +33,16 @@ namespace SalesProject.Api.Controllers
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [Route("api/[controller]")]
-        public IActionResult GetCustomers() =>
-            Ok(_customerRepository.GetAll());
+        public IActionResult GetCustomers()
+        {
+            var customers = _customerService.GetAll();
+
+            if (customers.Any())
+                return Ok(customers);
+
+            return NotFound($"Ops. Nenhum Cliente foi encontrado.");
+        }
+
 
         /// <summary>
         /// Get Customer by Id.
@@ -60,7 +59,7 @@ namespace SalesProject.Api.Controllers
         [Route("api/[controller]/{id:guid}")]
         public IActionResult GetCustomer(Guid id)
         {
-            var customer = _customerRepository.Get(id);
+            var customer = _customerService.Get(id);
 
             if (customer is not null)
                 return Ok(customer);
@@ -83,7 +82,7 @@ namespace SalesProject.Api.Controllers
         [Route("api/full/[controller]/{id:guid}")]
         public IActionResult GetFullCustomer(Guid id)
         {
-            var customer = _customerRepository.GetFullCustomer(id);
+            var customer = _customerService.GetFullCustomer(id);
 
             if (customer is not null)
                 return Ok(customer);
@@ -106,7 +105,7 @@ namespace SalesProject.Api.Controllers
         [Route("api/complete/[controller]/{id:guid}")]
         public IActionResult GetCompleteCustomer(Guid id)
         {
-            var customer = _customerRepository.GetCompleteCustomer(id);
+            var customer = _customerService.GetCustomerWithAdressesAndContacts(id);
 
             if (customer is not null)
                 return Ok(customer);
@@ -129,7 +128,7 @@ namespace SalesProject.Api.Controllers
         [Route("api/[controller]/name/{name}")]
         public IActionResult GetCustomersByName(string name)
         {
-            var customers = _customerRepository.GetByName(name);
+            var customers = _customerService.GetByName(name);
 
             if (customers.Count > 0)
                 return Ok(customers);
@@ -185,24 +184,14 @@ namespace SalesProject.Api.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var customer =
-                    new Customer(
-                        cnpj: RemoveCharacters(model.Cnpj),
-                        companyName: model.CompanyName,
-                        opening: DateTime.Parse(model.Opening).Date,
-                        phone: model.Phone,
-                        municipalRegistration: model.MunicipalRegistration,
-                        stateRegistration: model.StateRegistration,
-                        email: model.Email);
+            bool hasConflict = false;
+            var customer = _customerService.Create(model, ref hasConflict);
+
+            if (hasConflict)
+                return Conflict($"Ops. O cliente com Cnpj '{model.Cnpj}' já possuí um cadastro no sistema.");
 
             if (!customer.Valid)
                 return ValidationProblem(customer.GetAllNotifications());
-
-            if (_customerRepository.HasAnotherCustomerWithThisCnpj(customer.Cnpj))
-                return Conflict($"Ops. O cliente com Cnpj '{model.Cnpj}' já possuí um cadastro no sistema.");
-
-            _customerRepository.Create(customer);
-            _uow.Commit();
 
             return Created(
             $@"{HttpContext.Request.Scheme}://{HttpContext.Request.Host}{HttpContext.Request.Path}/{customer.Id}",
@@ -231,66 +220,14 @@ namespace SalesProject.Api.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var customer =
-                    new Customer(
-                        cnpj: RemoveCharacters(model.Cnpj),
-                        companyName: model.CompanyName,
-                        opening: DateTime.Parse(model.Opening).Date,
-                        phone: model.Phone,
-                        municipalRegistration: model.MunicipalRegistration,
-                        stateRegistration: model.StateRegistration,
-                        email: model.Email);
+            bool hasConflict = false;
+            var customer = _customerService.CreateCustomerWithAdressesAndContacts(model, ref hasConflict);
 
-            if (!customer.Valid)
-                return ValidationProblem($"{customer.GetNotification()}");
-
-            if (_customerRepository.HasAnotherCustomerWithThisCnpj(customer.Cnpj))
+            if (hasConflict)
                 return Conflict($"Ops. O cliente com Cnpj '{model.Cnpj}' já possuí um cadastro no sistema.");
 
-            foreach (var line in model.Adresses)
-            {
-                var address =
-                    new Address(
-                        description: line.Description,
-                        zipCode: line.ZipCode,
-                        type: (AddressType)line.Type,
-                        street: line.Street,
-                        neighborhood: line.Neighborhood,
-                        number: line.Number,
-                        city: line.City,
-                        state: line.State,
-                        customerId: customer.Id);
-
-                if (!address.Valid)
-                    return ValidationProblem($"{address.GetNotification()}");
-
-                address.SetCodeCity(codeCity: line.CodeCity);
-
-                customer.AddAddress(address);
-            }
-
-            if (!customer.Adresses.Where(a => a.Type == AddressType.Billing).Any())
-                return ValidationProblem($"Ops. Para adicionar um Cliente é necesssario um Endereço do tipo Cobrança");
-
-            foreach (var line in model.Contacts)
-            {
-                var contact =
-                    new Contact(firstName: line.FirstName,
-                                lastName: line.LastName,
-                                email: line.Email,
-                                whatsApp: line.WhatsApp,
-                                phone: line.Phone,
-                                customerId: customer.Id);
-
-                if (!contact.Valid)
-                    return ValidationProblem($"{contact.GetNotification()}");
-
-                customer.AddContact(contact);
-            }
-
-            _customerRepository.Create(customer);
-
-            _uow.Commit();
+            if (!customer.Valid)
+                return ValidationProblem(customer.GetAllNotifications());
 
             return Created(
             $@"{HttpContext.Request.Scheme}://{HttpContext.Request.Host}{HttpContext.Request.Path}/{customer.Id}",
@@ -311,15 +248,10 @@ namespace SalesProject.Api.Controllers
         [Route("api/[controller]/{id:guid}")]
         public IActionResult DeleteCustomer(Guid id)
         {
-            var customer = _customerRepository.Get(id);
+            if (_customerService.Delete(id))
+                return Ok();
 
-            if (customer is null)
-                return NotFound($"Ops. Cliente com Id:'{id}' não foi encontrado.");
-
-            _customerRepository.Delete(customer);
-            _uow.Commit();
-
-            return Ok();
+            return NotFound($"Ops. Cliente com Id: '{id}' não foi encontrado.");
         }
 
         /// <summary>
@@ -342,25 +274,15 @@ namespace SalesProject.Api.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var oldCustomer = _customerRepository.Get(id);
+            var updatedCustomer = _customerService.Edit(id, model);
 
-            if (oldCustomer is null)
-                return NotFound($"Ops. Cliente com Id:'{id}' não foi encontrado.");
+            if (updatedCustomer is null)
+                return NotFound($"Ops. Cliente com Id: '{id}' não foi encontrado.");
 
-            var newCustomer = oldCustomer.
-                        Edit(
-                        phone: model.Phone,
-                        municipalRegistration: model.MunicipalRegistration,
-                        stateRegistration: model.StateRegistration,
-                        email: model.Email);
+            if (!updatedCustomer.Valid)
+                return ValidationProblem(updatedCustomer.GetAllNotifications());
 
-            if (!newCustomer.Valid)
-                return ValidationProblem($"{newCustomer.GetNotification()}");
-
-            _customerRepository.Update(newCustomer);
-            _uow.Commit();
-
-            return Ok(newCustomer);
+            return Ok(updatedCustomer);
         }
 
         /// <summary>
@@ -383,201 +305,15 @@ namespace SalesProject.Api.Controllers
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var oldCustomer = _customerRepository.GetCompleteCustomer(id);
+            var updatedCustomer = _customerService.EditCustomerWithAdressesAndContacts(id, model);
 
-            if (oldCustomer is null)
-                return NotFound($"Ops. Cliente com Id:'{id}' não foi encontrado.");
+            if (updatedCustomer is null)
+                return NotFound($"Ops. Cliente com Id: '{id}' não foi encontrado.");
 
-            var newCustomer = oldCustomer.
-                        Edit(
-                        phone: model.Phone,
-                        municipalRegistration: model.MunicipalRegistration,
-                        stateRegistration: model.StateRegistration,
-                        email: model.Email);
+            if (!updatedCustomer.Valid)
+                return ValidationProblem(updatedCustomer.GetAllNotifications());
 
-            if (!newCustomer.Valid)
-                return ValidationProblem($"{newCustomer.GetNotification()}");
-
-            #region Manage Adresses
-
-            List<Address> newAdresses = new List<Address>();
-            List<Address> removedAdresses = new List<Address>();
-
-            foreach (var address in oldCustomer.Adresses)
-            {
-                bool remove = false;
-
-                for (int i = 0; i < model.Adresses.Count; i++)
-                {
-                    Guid currentIdAddress = string.IsNullOrEmpty(model.Adresses[i].Id)
-                    ? new Guid()
-                    : Guid.Parse(model.Adresses[i].Id);
-
-                    if (address.Id == currentIdAddress)
-                        break;
-
-                    if (i == model.Adresses.Count - 1)
-                    {
-                        remove = true;
-                        break;
-                    }
-                }
-                if (remove)
-                    removedAdresses.Add(address);
-            }
-
-            foreach (var addressModel in model.Adresses)
-            {
-                if (string.IsNullOrEmpty(addressModel.Id))
-                {
-                    var newAddress =
-                            new Address(
-                                description: addressModel.Description,
-                                zipCode: addressModel.ZipCode,
-                                type: (AddressType)addressModel.Type,
-                                street: addressModel.Street,
-                                neighborhood: addressModel.Neighborhood,
-                                number: addressModel.Number,
-                                city: addressModel.City,
-                                state: addressModel.State,
-                                customerId: oldCustomer.Id);
-
-                    if (!newAddress.Valid)
-                        return ValidationProblem($"{newAddress.GetNotification()}");
-
-                    newAddress.SetCodeCity(codeCity: addressModel.CodeCity);
-
-                    newAdresses.Add(newAddress);
-                }
-            }
-
-            foreach (var addressModel in model.Adresses)
-            {
-                if (string.IsNullOrEmpty(addressModel.Id))
-                    continue;
-
-                var address =
-                    newCustomer
-                    .Adresses
-                    .Where(c => c.Id == Guid.Parse(addressModel.Id))
-                    .FirstOrDefault();
-
-                address.Edit(
-                            description: addressModel.Description,
-                            zipCode: addressModel.ZipCode,
-                            type: (AddressType)addressModel.Type,
-                            street: addressModel.Street,
-                            neighborhood: addressModel.Neighborhood,
-                            number: addressModel.Number,
-                            city: addressModel.City,
-                            state: addressModel.State);
-
-                if (!address.Valid)
-                    return ValidationProblem($"{address.GetNotification()}");
-
-                address.SetCodeCity(codeCity: addressModel.CodeCity);
-            }
-
-            #endregion
-
-            #region Manage Contacts
-
-            List<Contact> newContacts = new List<Contact>();
-            List<Contact> removedContacts = new List<Contact>();
-
-            foreach (var contact in oldCustomer.Contacts)
-            {
-                bool remove = false;
-
-                for (int i = 0; i < model.Contacts.Count; i++)
-                {
-                    Guid currentIdContact = string.IsNullOrEmpty(model.Contacts[i].Id)
-                    ? new Guid()
-                    : Guid.Parse(model.Contacts[i].Id);
-
-                    if (contact.Id == currentIdContact)
-                        break;
-
-                    if (i == model.Contacts.Count - 1)
-                    {
-                        remove = true;
-                        break;
-                    }
-                }
-                if (remove)
-                    removedContacts.Add(contact);
-            }
-
-            foreach (var contactModel in model.Contacts)
-            {
-                if (string.IsNullOrEmpty(contactModel.Id))
-                {
-                    var newContact =
-                        new Contact(
-                            firstName: contactModel.FirstName,
-                            lastName: contactModel.LastName,
-                            email: contactModel.Email,
-                            whatsApp: contactModel.WhatsApp,
-                            phone: contactModel.Phone,
-                            customerId: oldCustomer.Id);
-
-                    if (!newContact.Valid)
-                        return ValidationProblem($"{newContact.GetNotification()}");
-
-                    newContacts.Add(newContact);
-                }
-            }
-
-            foreach (var contactModel in model.Contacts)
-            {
-                if (string.IsNullOrEmpty(contactModel.Id))
-                    continue;
-
-                var contact =
-                    newCustomer
-                    .Contacts
-                    .Where(c => c.Id == Guid.Parse(contactModel.Id))
-                    .FirstOrDefault();
-
-                contact.Edit(
-                        firstName: contactModel.FirstName,
-                        lastName: contactModel.LastName,
-                        email: contactModel.Email,
-                        whatsApp: contactModel.WhatsApp,
-                        phone: contactModel.Phone);
-
-                if (!contact.Valid)
-                    return ValidationProblem($"{contact.GetNotification()}");
-            }
-
-            #endregion
-
-            if (newAdresses.Any())
-                foreach (var newAddress in newAdresses)
-                    newCustomer.AddAddress(newAddress);
-
-            if (newContacts.Any())
-                foreach (var contact in newContacts)
-                    newCustomer.AddContact(contact);
-
-            if (removedAdresses.Any())
-                foreach (var address in removedAdresses)
-                    newCustomer.RemoveAddress(address);
-
-            if (removedContacts.Any())
-                foreach (var contact in removedContacts)
-                    newCustomer.RemoveContact(contact);
-
-            if (!newCustomer.Adresses.Where(a => a.Type == AddressType.Billing).Any())
-                return ValidationProblem($"Ops. Falha ao editar Cliente '{newCustomer.CompanyName}', deve possuir um Endereço do tipo Cobrança");
-
-            _customerRepository.Update(newCustomer);
-            _uow.Commit();
-
-            return Ok(newCustomer);
+            return Ok(updatedCustomer);
         }
-
-        private string RemoveCharacters(string cnpj) =>
-            cnpj.Replace(".", string.Empty).Replace("/", string.Empty).Replace("-", string.Empty);
     }
 }
