+++ SalesProject.Domain/Interfaces/Repository/ICustomerRepository.cs
@@ -6,11 +6,11 @@ namespace SalesProject.Domain.Interfaces.Repository
 {
     public interface ICustomerRepository : IDisposable
     {
-        List<Customer> GetAll();
+        ICollection<Customer> GetAll();
         Customer Get(Guid id);
         Customer GetFullCustomer(Guid id);
-        Customer GetCompleteCustomer(Guid id);
-        List<Customer> GetByName(string name);
+        Customer GetCustomerWithAdressesAndContacts(Guid id);
+        ICollection<Customer> GetByName(string name);
         bool HasAnotherCustomerWithThisCnpj(string cnpj);
         void Create(Customer customer);
         void Update(Customer customer);
